% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_interval.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate quantiles of a tidy dataframe}
\usage{
calc_quantiles(data, intervals)
}
\arguments{
\item{data}{A data frame with containing x and y columns, with several y
values for every x}

\item{intervals}{a list of intervals for which corresponding quantiles are
desired.}
}
\value{
A data frame containing x, y, and quantile columns.
}
\description{
Calculate quantiles of a tidy dataframe
}
\examples{

head(fake_df)

fake_q <- calc_quantiles(fake_df, intervals=c(0,0.5,0.8))
head(fake_q)

}
