% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_interval.R
\name{.in_numeric}
\alias{.in_numeric}
\title{Find elements of one numeric vector in another.}
\usage{
.in_numeric(a, b, tol = 1e-06)
}
\arguments{
\item{a}{The vector for which comparison are to be made}

\item{b}{The vector to match against.}

\item{tol}{The tolerance within which values are assumed to be the same.}
}
\value{
A vector of logical values the same length as `a` describing whether
this vector element is close to at least one element in `b`
}
\description{
Find which elements of `a` are equal to at least one element in `b`, to
within some tolerance tol.
}
